Compiler warning fixes... but
Here is where I gave up on gtkmm 2.x.

Going to do a rewrite and then might as well use gtkmm 3.24.

+++ PgnDatabase.cc
@@ -214,7 +214,7 @@ class Scanner {
 
     int push_position(void)
     {
-      size_t size = M_stack.size();
+      int size = M_stack.size();
       if (G_UNLIKELY(size <= M_stack_index))
 	M_stack.push_back(M_current_position);
       else
